# apheresis-openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://apheresis-plasma-service.api.svc.cluster.local","description":"Generated server url"}],"tags":[{"name":"Actuator","description":"Monitor and interact","externalDocs":{"description":"Spring Boot Actuator Web API Documentation","url":"https://docs.spring.io/spring-boot/docs/current/actuator-api/html/"}}],"paths":{"/management/loggers/{name}":{"get":{"tags":["Actuator"],"summary":"Actuator web endpoint 'loggers-name'","operationId":"loggerLevels","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/vnd.spring-boot.actuator.v3+json":{"schema":{"type":"object"}},"application/vnd.spring-boot.actuator.v2+json":{"schema":{"type":"object"}},"application/json":{"schema":{"type":"object"}}}},"404":{"description":"Not Found"}}},"post":{"tags":["Actuator"],"summary":"Actuator web endpoint 'loggers-name'","operationId":"configureLogLevel","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"string","enum":["TRACE","DEBUG","INFO","WARN","ERROR","FATAL","OFF"]}}}},"responses":{"204":{"description":"No Content"},"400":{"description":"Bad Request"}}}},"/springwolf/docs.yaml":{"get":{"tags":["async-api-controller"],"operationId":"asyncApiYaml","parameters":[],"responses":{"200":{"description":"OK","content":{"application/yaml":{"schema":{"type":"string"}}}}}}},"/springwolf/docs":{"get":{"tags":["async-api-controller"],"operationId":"asyncApiYaml_1","parameters":[],"responses":{"200":{"description":"OK","content":{"application/yaml":{"schema":{"type":"string"}},"application/json":{"schema":{"type":"string"}}}}}}},"/springwolf/docs/{group}":{"get":{"tags":["async-api-controller"],"operationId":"asyncApiYaml_2","parameters":[{"name":"group","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/yaml":{"schema":{"type":"string"}},"application/json":{"schema":{"type":"string"}}}}}}},"/springwolf/docs.json":{"get":{"tags":["async-api-controller"],"operationId":"asyncApiJson","parameters":[],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/springwolf/ui-config":{"get":{"tags":["ui-config-controller"],"operationId":"getUiConfig","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/io.github.springwolf.core.controller.UiConfigController.UiConfig"}}}}}}},"/management/threaddump":{"get":{"tags":["Actuator"],"summary":"Actuator web endpoint 'threaddump'","operationId":"threadDump","responses":{"200":{"description":"OK","content":{"text/plain;charset=UTF-8":{"schema":{"type":"object"}},"application/vnd.spring-boot.actuator.v3+json":{"schema":{"type":"object"}},"application/vnd.spring-boot.actuator.v2+json":{"schema":{"type":"object"}},"application/json":{"schema":{"type":"object"}}}}}}},"/management/loggers":{"get":{"tags":["Actuator"],"summary":"Actuator web endpoint 'loggers'","operationId":"loggers","responses":{"200":{"description":"OK","content":{"application/vnd.spring-boot.actuator.v3+json":{"schema":{"type":"object"}},"application/vnd.spring-boot.actuator.v2+json":{"schema":{"type":"object"}},"application/json":{"schema":{"type":"object"}}}}}}},"/management/info":{"get":{"tags":["Actuator"],"summary":"Actuator web endpoint 'info'","operationId":"info","responses":{"200":{"description":"OK","content":{"application/vnd.spring-boot.actuator.v3+json":{"schema":{"type":"object"}},"application/vnd.spring-boot.actuator.v2+json":{"schema":{"type":"object"}},"application/json":{"schema":{"type":"object"}}}}}}},"/management/health/{path}":{"get":{"tags":["Actuator"],"summary":"Actuator web endpoint 'health-path'","operationId":"health","parameters":[{"name":"path","in":"path","required":true,"schema":{"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"OK","content":{"application/vnd.spring-boot.actuator.v3+json":{"schema":{"type":"object"}},"application/vnd.spring-boot.actuator.v2+json":{"schema":{"type":"object"}},"application/json":{"schema":{"type":"object"}}}},"404":{"description":"Not Found"}}}},"/management/health":{"get":{"tags":["Actuator"],"summary":"Actuator web endpoint 'health'","operationId":"health_1","responses":{"200":{"description":"OK","content":{"application/vnd.spring-boot.actuator.v3+json":{"schema":{"type":"object"}},"application/vnd.spring-boot.actuator.v2+json":{"schema":{"type":"object"}},"application/json":{"schema":{"type":"object"}}}}}}},"/management/env/{toMatch}":{"get":{"tags":["Actuator"],"summary":"Actuator web endpoint 'env-toMatch'","operationId":"environmentEntry","parameters":[{"name":"toMatch","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/vnd.spring-boot.actuator.v3+json":{"schema":{"type":"object"}},"application/vnd.spring-boot.actuator.v2+json":{"schema":{"type":"object"}},"application/json":{"schema":{"type":"object"}}}},"404":{"description":"Not Found"}}}},"/management/env":{"get":{"tags":["Actuator"],"summary":"Actuator web endpoint 'env'","operationId":"environment","parameters":[{"name":"pattern","in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/vnd.spring-boot.actuator.v3+json":{"schema":{"type":"object"}},"application/vnd.spring-boot.actuator.v2+json":{"schema":{"type":"object"}},"application/json":{"schema":{"type":"object"}}}}}}},"/management/configprops/{prefix}":{"get":{"tags":["Actuator"],"summary":"Actuator web endpoint 'configprops-prefix'","operationId":"configurationPropertiesWithPrefix","parameters":[{"name":"prefix","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/vnd.spring-boot.actuator.v3+json":{"schema":{"type":"object"}},"application/vnd.spring-boot.actuator.v2+json":{"schema":{"type":"object"}},"application/json":{"schema":{"type":"object"}}}},"404":{"description":"Not Found"}}}},"/management/configprops":{"get":{"tags":["Actuator"],"summary":"Actuator web endpoint 'configprops'","operationId":"configurationProperties","responses":{"200":{"description":"OK","content":{"application/vnd.spring-boot.actuator.v3+json":{"schema":{"type":"object"}},"application/vnd.spring-boot.actuator.v2+json":{"schema":{"type":"object"}},"application/json":{"schema":{"type":"object"}}}}}}},"/management":{"get":{"tags":["Actuator"],"summary":"Actuator root web endpoint","operationId":"links","responses":{"200":{"description":"OK","content":{"application/vnd.spring-boot.actuator.v3+json":{"schema":{"type":"object","additionalProperties":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/org.springframework.boot.actuate.endpoint.web.Link"}}}},"application/vnd.spring-boot.actuator.v2+json":{"schema":{"type":"object","additionalProperties":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/org.springframework.boot.actuate.endpoint.web.Link"}}}},"application/json":{"schema":{"type":"object","additionalProperties":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/org.springframework.boot.actuate.endpoint.web.Link"}}}}}}}}}},"components":{"schemas":{"io.github.springwolf.core.controller.UiConfigController.UiConfig":{"type":"object","properties":{"initialConfig":{"$ref":"#/components/schemas/io.github.springwolf.core.controller.UiConfigController.UiConfig.InitialConfig"},"groups":{"type":"array","items":{"$ref":"#/components/schemas/io.github.springwolf.core.controller.UiConfigController.UiConfig.UiConfigGroup"}}},"title":"UiConfig"},"io.github.springwolf.core.controller.UiConfigController.UiConfig.InitialConfig":{"type":"object","properties":{"showBindings":{"type":"boolean"},"showHeaders":{"type":"boolean"}},"title":"InitialConfig"},"io.github.springwolf.core.controller.UiConfigController.UiConfig.UiConfigGroup":{"type":"object","properties":{"name":{"type":"string"}},"title":"UiConfigGroup"},"org.springframework.boot.actuate.endpoint.web.Link":{"type":"object","properties":{"href":{"type":"string"},"templated":{"type":"boolean"}},"title":"Link"}}}}